+++ js/modules/mixed-air/state.js
@@ -1,0 +1,12 @@
+import { createModuleState } from '../../core/state.js';
+
+export const state = createModuleState({
+  mixingOutdoorVolumeFlowM3h: '',
+  mixingOutdoorTemp: '',
+  mixingOutdoorRh: '',
+  mixingRecircVolumeFlowM3h: '',
+  mixingRecircTemp: '',
+  mixingRecircRh: ''
+});
+
+export default state;